add NullDataCompressor fix CompressDataSerilizer => CompressDataSerializer
