+++ samples/DotNetCoreSample/DotNetCoreSample.csproj
@@ -13,7 +13,7 @@
     <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="2.0.0" />
     <PackageReference Include="Microsoft.Extensions.ObjectPool" Version="2.2.0" />
     <PackageReference Include="System.Data.SqlClient" Version="4.4.0" />
-    <PackageReference Include="WeihanLi.Common" Version="1.0.15" />
+    <PackageReference Include="WeihanLi.Common" Version="1.0.16" />
     <PackageReference Include="WeihanLi.Common.Logging.Log4net" Version="1.0.13" />
   </ItemGroup>
