+++ src/WeihanLi.Common/Helpers/DataSerializer.cs
@@ -27,63 +27,6 @@ namespace WeihanLi.Common.Helpers
         T Deserializer<T>([NotNull]byte[] bytes);
     }
 
-    /// <summary>
-    /// DataCompressor
-    /// </summary>
-    public interface IDataCompressor
-    {
-        /// <summary>
-        /// compress data
-        /// </summary>
-        /// <param name="sourceData">source data</param>
-        /// <returns>copmpressed data</returns>
-        byte[] Compress([NotNull]byte[] sourceData);
-
-        /// <summary>
-        /// compress data async
-        /// </summary>
-        /// <param name="sourceData">source data</param>
-        /// <returns>copmpressed data</returns>
-        Task<byte[]> CompressAsync([NotNull]byte[] sourceData);
-
-        /// <summary>
-        /// decompress compressed data
-        /// </summary>
-        /// <param name="compressedData">copmpressed data</param>
-        /// <returns>source data</returns>
-        byte[] Decompress([NotNull]byte[] compressedData);
-
-        /// <summary>
-        /// decompress compressed data async
-        /// </summary>
-        /// <param name="compressedData">copmpressed data</param>
-        /// <returns>source data</returns>
-        Task<byte[]> DecompressAsync([NotNull]byte[] compressedData);
-    }
-
-    public class GZipDataCompressor : IDataCompressor
-    {
-        public byte[] Compress([NotNull]byte[] sourceData)
-        {
-            return sourceData.CompressGZip();
-        }
-
-        public Task<byte[]> CompressAsync([NotNull]byte[] sourceData)
-        {
-            return sourceData.CompressGZipAsync();
-        }
-
-        public byte[] Decompress([NotNull]byte[] compressedData)
-        {
-            return compressedData.DecompressGZip();
-        }
-
-        public Task<byte[]> DecompressAsync([NotNull]byte[] compressedData)
-        {
-            return compressedData.DecompressGZipAsync();
-        }
-    }
-
     public class BinaryDataSerializer : IDataSerializer
     {
         private readonly BinaryFormatter _binaryFormatter = new BinaryFormatter();
@@ -165,32 +108,12 @@ namespace WeihanLi.Common.Helpers
         }
     }
 
-    public class CompressGZipSerilizer : IDataSerializer
-    {
-        private readonly IDataSerializer _serializer;
-
-        public CompressGZipSerilizer(IDataSerializer serializer)
-        {
-            _serializer = serializer;
-        }
-
-        public byte[] Serialize<T>(T obj)
-        {
-            return _serializer.Serialize(obj).CompressGZip();
-        }
-
-        public T Deserializer<T>(byte[] bytes)
-        {
-            return _serializer.Deserializer<T>(bytes.DecompressGZip());
-        }
-    }
-
-    public class CompressDataSerilizer : IDataSerializer
+    public class CompressDataSerializer : IDataSerializer
     {
         private readonly IDataSerializer _serializer;
         private readonly IDataCompressor _compressor;
 
-        public CompressDataSerilizer(IDataSerializer serializer, IDataCompressor compressor)
+        public CompressDataSerializer(IDataSerializer serializer, IDataCompressor compressor)
         {
             _serializer = serializer;
             _compressor = compressor;
